+++ include/ast/exprs/struct_instance.h
@@ -1,30 +1,31 @@
 #pragma once
 #include <ast/expr.h>
 #include <basic/name.h>
+#include <basic/types/type.h>
 #include <tuple>
 #include <vector>
 
 namespace veo::ast {
 
 class StructInstance : public Expr {
-    basic::NameObj                                  _path;
+    basic::Type                                    *_structType;
     std::vector<std::tuple<basic::NameObj, Expr *>> _fields;
 
 public:
     StructInstance (
-        basic::NameObj                                  path,
+        basic::Type                                    *structType,
         std::vector<std::tuple<basic::NameObj, Expr *>> fields,
         llvm::SMLoc                                     start,
         llvm::SMLoc                                     end)
-        : _path (std::move (path)),
+        : _structType (structType),
           _fields (std::move (fields)),
           Expr (NodeKind::StructInstance, start, end) {}
 
     ast_classof (StructInstance);
 
-    basic::NameObj
-    Path () const {
-        return _path;
+    basic::Type *&
+    StructType () {
+        return _structType;
     }
 
     std::vector<std::tuple<basic::NameObj, Expr *>> &