+++ include/basic/types/all.h
@@ -1,3 +1,4 @@
+#include <basic/types/alias.h>
 #include <basic/types/bool.h>
 #include <basic/types/char.h>
 #include <basic/types/floating.h>
@@ -8,4 +9,5 @@
 #include <basic/types/size.h>
 #include <basic/types/struct.h>
 #include <basic/types/trait.h>
+#include <basic/types/trait_this.h>
 #include <basic/types/type.h>