+++ include/basic/types/type.h
@@ -14,6 +14,8 @@ class TraitType;
 class NamedType;
 class PointerType;
 class NothType;
+class AliasType;
+class TraitThisType;
 
 #define type_classof(kind)                                                               \
     static bool classof (const veo::basic::Type *type) {                                 \
@@ -30,7 +32,9 @@ enum class TypeKind : uint8_t {
     Trait,
     Named,
     Pointer,
-    Noth
+    Noth,
+    Alias,
+    TraitThis,
 };
 
 class Type {
@@ -80,6 +84,10 @@ public:
     as (Pointer);
     is (Noth);
     as (Noth);
+    is (Alias);
+    as (Alias);
+    is (TraitThis);
+    as (TraitThis);
 
     bool
     IsIntOrSize () const {
@@ -90,6 +98,12 @@ public:
     IsNumber () const {
         return IsInteger () || IsFloating () || IsSize ();
     }
+
+    const Type *
+    CanonicalType () const;
+
+    Type *
+    CanonicalType ();
 
 #undef as
 #undef is